feat(convex-native): Phase 0 — native read-layer client plumbing
Phase 0 of the Convex-native read layer migration (docs/designs/convex-native-read-layer.md).
Additive only — no surface reads through the new plumbing yet, so behaviour is unchanged.

- Add convex-helpers@0.1.119 and wrap the app in ConvexQueryCacheProvider inside
  the existing ConvexProviderWithAuth tree (bounded: 5min expiration, 250 idle
  entries). Affects only call sites that opt into the cached hooks from
  "convex-helpers/react/cache"; the default convex/react useQuery is untouched.
- Make the RBAC logic isomorphic: move the pure permission core (RESOURCES,
  rolePermissions, hasPermission, isReadOnly, role helpers + types) to
  convex/lib/permissions-core.ts so Convex functions and the Next server share one
  source of truth. src/lib/permissions.ts re-exports it and keeps the UI-only
  PERMISSION_REGISTRY + roleLabels. No behaviour change.
- Land the migration design doc + independent review response.

Verification:
- npx tsc --noEmit → exit 0 (after prisma generate)
- pnpm run lint → exit 0 (0 errors, pre-existing warnings only)
- pnpm exec vitest run -t permission → 22 passed
- pnpm install --frozen-lockfile → consistent (convex-helpers 0.1.119 locked)
