+++ package.json
@@ -108,6 +108,7 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "convex": "^1.40.0",
+    "convex-helpers": "^0.1.119",
     "date-fns": "^4.4.0",
     "file-type": "^22.0.1",
     "framer-motion": "^12.40.0",
